+++ slicer/bin/args.rs
@@ -3,7 +3,7 @@ use std::{any::Any, path::PathBuf, str::FromStr};
 use anyhow::{Context, Ok, Result};
 use clap::{ArgMatches, Parser};
 use common::{
-    slice::{ExposureConfig, Format, SliceConfig},
+    slice::{ExposureConfig, SliceConfig, SliceMode},
     units::{Milimeters, MilimetersPerMinute, Seconds},
 };
 use nalgebra::{ArrayStorage, Const, Matrix, Scalar, U1, Vector2, Vector3};
@@ -111,11 +111,9 @@ pub struct Model {
 }
 
 impl Args {
-    pub fn slice_config(&self, extension: &str) -> Result<SliceConfig> {
-        let format = Format::from_extension(extension).context("Unknown output format")?;
-
+    pub fn slice_config(&self) -> Result<SliceConfig> {
         Ok(SliceConfig {
-            format,
+            mode: SliceMode::Raster,
             supersample: cbrt(self.supersample) as u8,
             platform_resolution: self.platform_resolution,
             platform_size: self.platform_size.map(Milimeters::new),